Refactor and update logic
This moves the logic around a bit. First I've split the old
(score_method) and new (rule_method) methods apart since the code was
getting a bit long.

Secondly, I have updated the logic of the new method to handle ribozymes
as well. This will now do a better job of picking the rna_type and
hopefully name given ribozymes. In addition, it will ignore the generic
annotations like 'misc_RNA' and 'other' if possible.
